+++ controllers/examples/social_image.py
@@ -6,7 +6,7 @@ Generate the image used for the GitHub social preview card.
 """
 
 if __name__ == "__main__":
-    m = Magnebot(launch_build=False, screen_width=1024, screen_height=1024)
+    m = Magnebot(launch_build=False, screen_width=1024, screen_height=1024, skip_frames=0)
     m.init_scene(scene="2b", layout=1, room=1)
 
     # Add a third-person camera.