+++ controllers/tests/turn_constants/turn_constants.py
@@ -11,32 +11,48 @@ Calculate the time elapsed for a single turn action given a range of combination
 if __name__ == "__main__":
     m = TestController()
     m._debug = False
+    # Destroy the old data.
     p = Path("../data/turn_constants.csv")
     if p.exists():
         p.unlink()
-    header = "time,magic_number,outer_track,front\n"
+    # Write the header of the spreadsheet to disk.
+    header = "time,magic_number,outer_track,front,drift\n"
     p.write_text(header)
     path = str(p.resolve())
+
+    origin = np.array([0, 0, 0])
+    # Iterate through many combinations of the turn constants.
     for magic_number in np.arange(start=2.4, stop=3.4, step=0.1):
-        for outer_track in np.arange(start=1, stop=1.6, step=0.1):
+        for outer_track in np.arange(start=1, stop=2.1, step=0.1):
             for front in np.arange(start=0.5, stop=1.5, step=0.1):
                 # Set the turn constants.
                 Magnebot._TURN_MAGIC_NUMBER = magic_number
                 Magnebot._TURN_OUTER_TRACK = outer_track
                 Magnebot._TURN_FRONT = front
                 m.init_scene()
-                # Turn the Magnebot. Record how long the turn takes.
+                # Turn the Magnebot.
                 t0 = time()
-                status = m.turn_by(45)
+                status = m.turn_by(120)
+                # Ignore the result if the action failed.
                 if status != ActionStatus.success:
-                    t1 = 1000
+                    continue
                 else:
-                    t1 = time() - t0
-                # Write the result to disk.
-                line = f"{t1},{round(magic_number, 3)},{round(outer_track, 3)},{round(front, 3)}"
-                print(line)
-                with open(path, "at") as f:
-                    f.write(line + "\n")
+                    magnebot_position = m.state.magnebot_transform.position
+                    magnebot_position[1] = 0
+                    # Ignore the result if the Magnebot drifted too far away from the origin.
+                    drift = np.linalg.norm(magnebot_position - origin)
+                    if drift > 0.025:
+                        continue
+                    # Success!
+                    else:
+                        # Get the time elapsed.
+                        t1 = time() - t0
+                        # Write the result to disk.
+                        line = f"{t1},{round(magic_number, 3)},{round(outer_track, 3)},{round(front, 3)}," \
+                               f"{round(drift, 3)}"
+                        print(line)
+                        with open(path, "at") as f:
+                            f.write(line + "\n")
     m.end()
 
     # Print the shortest time.